app/src/main/java/com/example/kotlin_coroutines_and_flow/usecases/coroutines/usecase1/PerformSingleNetworkRequestViewModel.kt
package com.example.kotlin_coroutines_and_flow.usecases.coroutines.usecase1

import androidx.lifecycle.viewModelScope
import com.example.kotlin_coroutines_and_flow.base.BaseViewModel
import com.example.kotlin_coroutines_and_flow.mock.MockApi
import kotlinx.coroutines.launch
import timber.log.Timber

class PerformSingleNetworkRequestViewModel(
    private val mockApi: MockApi = mockApi()
) : BaseViewModel<UiState>() {

    fun performSingleNetworkRequest() {
        uiState.value = UiState.Loading
        viewModelScope.launch {
            try {
                val recentAndroidVersions = mockApi.getRecentAndroidVersions()
                uiState.value = UiState.Success(recentAndroidVersions)
            } catch (exception: Exception) {
                Timber.e(exception)
                uiState.value = UiState.Error("Network Request failed!")
            }
        }
    }
}